actualizacion de registro alimento

+++ lib/controller/con_actualizar_alimento.dart
@@ -22,15 +22,14 @@ class con_actualizar_alimento {
       String porcion) async {
     final alimento = model_alimento();
     final id_al_int = int.parse(id_al);
-    final calorias_int = calorias.isNotEmpty ? double.parse(calorias) : 0;
-    final azuccares_int = azuccares.isNotEmpty ? double.parse(azuccares) : 0;
-    final proteina_int = proteina.isNotEmpty ? double.parse(proteina) : 0;
-    final sodio_int = sodio.isNotEmpty ? double.parse(sodio) : 0;
-    final grasa_total_int =
-        grasa_total.isNotEmpty ? double.parse(grasa_total) : 0;
-    final h_de_c_int = h_de_c.isNotEmpty ? double.parse(h_de_c) : 0;
-    final colesterol_int = colesterol.isNotEmpty ? double.parse(colesterol) : 0;
-    final porcion_int = porcion.isNotEmpty ? double.parse(porcion) : 0;
+    final calorias_int = calorias.isNotEmpty ? int.parse(calorias) : 0;
+    final azuccares_int = azuccares.isNotEmpty ? int.parse(azuccares) : 0;
+    final proteina_int = proteina.isNotEmpty ? int.parse(proteina) : 0;
+    final sodio_int = sodio.isNotEmpty ? int.parse(sodio) : 0;
+    final grasa_total_int = grasa_total.isNotEmpty ? int.parse(grasa_total) : 0;
+    final h_de_c_int = h_de_c.isNotEmpty ? int.parse(h_de_c) : 0;
+    final colesterol_int = colesterol.isNotEmpty ? int.parse(colesterol) : 0;
+    final porcion_int = porcion.isNotEmpty ? int.parse(porcion) : 0;
     final act_alimento = alimento.actualizar_alimento(
         id_al_int,
         nombre,
